requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
djangorestframework==3.9.2
drf-yasg==1.15.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
